pyproject.toml
[tool.poetry]
name = "msft-stock-agent"
version = "0.1.0"
description = ""
authors = ["kimtth <kimtth@email.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
ta-lib = {path = "config/TA_Lib-0.4.32-cp311-cp311-win_amd64.whl"}
quantstats = "^0.0.62"
pyautogen = "^0.2.35"
autogenstudio = "^0.1.5"
yfinance = "^0.2.43"
pandas = "^2.2.2"
zipline-reloaded = "^3.0.4"
docker = "^7.1.0"
pyyaml = "^6.0.2"
ta = "^0.11.0"
python-dotenv = "^1.0.1"
openpyxl = "^3.1.5"
click = "^8.1.7"
beautifulsoup4 = "^4.12.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
